数据类型、引入编码

+++ backend/ruoyi-system/src/main/java/com/ruoyi/system/domain/IcesMaterialStock.java
@@ -42,7 +42,7 @@ public class IcesMaterialStock extends BaseEntity {
     /**
      * 库存
      */
-    private Long mssStock;
+    private Float mssStock;
     /**
      * 已删除
      */
