+++ backend/ruoyi-system/src/main/java/com/ruoyi/system/domain/vo/IcesMaterialStockVo.java
@@ -49,7 +49,7 @@ public class IcesMaterialStockVo {
      * 库存
      */
     @ExcelProperty(value = "库存")
-    private Long mssStock;
+    private Float mssStock;
 
     /**
      * 已删除
